Allow user input pameters whose number is variable

+++ contacts.py
@@ -1,47 +1,40 @@
 import sys
+import os
 import pickle as p
-
-class contacts:
-	def __init__(self, name, telephone, address):
-		self.name = name
-		self.telephone = telephone
-		self.address = address
-		
 
 if len(sys.argv) < 2:
 	print("No action!")
 	sys.exit()
 
-filename = "contacts.txt"
+filename = "contractor.txt"
 action = sys.argv[1]
+dict = {}
 print(action)
 if action == "add":	
 	name = sys.argv[2][:]
-	telephone = sys.argv[3][:]
-	address = sys.argv[4][:]
-	# print(name)
-	# print(telephone)
-	# print(address)
-	# person = contacts(name, telephone, address)
-	fd = open(filename, "rb")
-	dict = p.load(fd)
-	print(dict)
-	fd.close()
-	
-	dict[name] = [telephone, address]
-	fd = open(filename, "wb")	
-	p.dump(dict, fd)
-	fd.close()
+	info = sys.argv[3:]
+	if os.path.isfile(filename):
+		fd = open(filename, "rb")
+		dict = p.load(fd)
+		fd.close()
+		
+		dict[name] = info
+		fd = open(filename, "wb")	
+		p.dump(dict, fd)
+		fd.close()
+	else:
+		dict[name] = info
+		fd = open(filename, "wb")	
+		p.dump(dict, fd)
+		fd.close()
 	
 elif action == "del":
-	# print(action)	
 	name = sys.argv[2][:]
 	fd = open(filename, "rb")
 	dict = p.load(fd)
 	del dict[name]
 	fd.close()
-	
-	# dict[name] = [telephone, address]
+
 	fd = open(filename, "wb")	
 	p.dump(dict, fd)
 	fd.close()
